+++ app/routes/tasks/no-tasks.tsx
@@ -1,0 +1,10 @@
+import { Link } from "@remix-run/react";
+
+export const NoTasks: React.FC = () => (
+  <li className="task">
+    <p>No tasks found.</p>
+    <p>
+      <Link to="/add-task">Add one!</Link>
+    </p>
+  </li>
+);
