+++ app/types/index.ts
@@ -1,0 +1,9 @@
+interface Task {
+  id: string;
+  title: string;
+  content: string;
+}
+
+interface TaskListProps {
+  tasks: Task[];
+}
